fix cs_prop() issue #127
The problem was with the EPIsuite data.  Some compounds don’t have
values for all types of data so occasionally results from gsub() were
vectors of zero length.  I added an if statement after every gsub to
check that something actually got added to the vector.  This is maybe
not the most elegant solution, and is probably overkill because not all
properties are subject to the possibility of missing values (I think
only experimental values are missing, not predicted values).