Added robustness to parsing of values that might be missing.
In the current code when a value is not found it returns `numeric(0)` which has length 0. SO when added to the vector things get out of sync.
Example that triggers the bug: http://www.chemspider.com/Chemical-Structure.391783.html

I didn't fix for all the properties but I think everything can be fixed the same way.